DatabaseTable find function extended
The find function can now either look for a single match against a column in a database table, or look for matches for multiple values on multiple columns. This was done to simplify the getUserPrediction method on the prediction entity to enable the prediction table to be searched for predictions for a particular userId and matchId.