Prediction edit page - when predictions are saved, the page redisplays
Rather than going back to the userMatchPredictions (Fixtures) page, the page just refreshes. Additionally, code was removed from the Prediction saveEdit method that allowed empty teams to be submitted for a prediction as this is no longer required.

Added flags, dateTime and stage to the prediction edit page, and rearranged elements of the page.